refactor(stella-serve): move the turn registry out of server.rs
server.rs was 1492 lines against the gate's strict 1500-line ceiling with
no baseline entry, so the next non-trivial change to it would have failed
make gate outright and check-file-size.sh --update would have refused to
grandfather a first-time crossing (#3441). Eight lines is not headroom.

The seam is the one the file's own module doc already drew: it is the
transport — listener, connection fold, auth, route classification — and
src/state.rs is what the transport serves out of. ServerState, the Entry
one live turn owns, and the registry's three rules move together because
they are one invariant: the MAX_LIVE_TURNS cap is a queue rather than a
one-way latch only because reclamation runs inside the same lock hold
that admits. src/routes.rs was carved out of the same file earlier and is
the precedent.

Two seams the move forced, both narrowing rather than widening: the state
is now built by ServerState::new(config, host_policy) instead of a struct
literal at the call site, which is also where the two config-proposed
ceilings are clamped; and the bearer token is compared through
token_matches rather than handed out, because a secret a caller holds is
one careless format! away from a log line.

No behaviour change and no baseline entry.

Closes #3734
