+++ crates/stella-serve/src/lib.rs
@@ -90,6 +90,7 @@ pub mod schema_export;
 mod server;
 mod session;
 mod sessions;
+mod state;
 pub mod subagents;
 mod throttle;
